Cycler Detection & Absolute Time
Adding the functionality to detect the cycler based on the keywords found and the option for the user to tell it or the line from where data starts and metadata finishes.

Implemented Absolute time detection taking it from the available recorded testing data or the metadata of the experiment/file.

Adjusted the logger for the new changes